requirements.txt
ultralytics
streamlit
gradio
fastapi
opencv-python
pillow
pyyaml
